requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
django-environ==0.11.2
django-extensions==3.2.3
Pillow==10.1.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
django-filter==23.5 